requirements.txt
Fabric
Flask-SQLAlchemy
Flask-Restful
Flask-Marshmallow
Marshmallow-SQLAlchemy
Flask-WTF
WTForms
coverage
# ecdsa
# paramiko
# pycrypto